Extract DangerAction to a common folder
